Excluded some files from .idea dir Removed __pycache__ and .pyc..new .gitignore file Moved modules into a dedicated directory called ArrayTools Added a test script in the root directory
