+++ .gitignore
@@ -1,0 +1,7 @@
+.idea/workspace.xml
+.idea/usage.statistics.xml
+.idea/tasks.xml
+.idea/misc.xml
+.idea/*.iml
+__pycache__
+*.pyc
