+++ ArrayTools/arrayClass.py
